requirements.txt
streamlit>=1.38.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0
transformers>=4.40.0
Pillow>=9.1.0
faiss-cpu>=1.8.0
pypdf>=5.0.0
python-dotenv>=1.0.0
tavily-python>=0.5.0